move configuration to lib
